refactor(ReplayView): remove unused Replayer class and workers
The Replayer class in the ReplayView utility folder and the local.worker.ts and
remote.worker.ts files have been removed as they are not utilized in the current
implementation. Additionally, redundant console.log statements and error handling
in the fetchZIP function have been cleaned up to improve code readability and
maintainability.

BREAKING CHANGE: The removal of the Replayer class and workers may affect any
dependent code or functionality that relies on these components.